# Use new storage.Appender Append() instead of Add & AddFast