--- a/pkg/query/test.go
+++ b/pkg/query/test.go
@@ -434,7 +434,7 @@ func (cmd *loadCmd) Append(a storage.Appender) error {
 		m := cmd.metrics[h]
 
 		for _, s := range smpls {
-			if _, err := a.Add(m, s.T, s.V); err != nil {
+			if _, err := a.Append(0, m, s.T, s.V); err != nil {
 				return err
 			}
 		}