--- a/pkg/receive/handler_test.go
+++ b/pkg/receive/handler_test.go
@@ -262,7 +262,7 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -273,7 +273,7 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -284,7 +284,7 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 			},
 		},
@@ -295,10 +295,10 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -309,13 +309,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -326,13 +326,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -343,13 +343,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -360,13 +360,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -377,15 +377,15 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
+					appender:    newFakeAppender(nil, nil, nil),
 					appenderErr: appenderErrFn,
 				},
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
+					appender:    newFakeAppender(nil, nil, nil),
 					appenderErr: appenderErrFn,
 				},
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
+					appender:    newFakeAppender(nil, nil, nil),
 					appenderErr: appenderErrFn,
 				},
 			},
@@ -397,13 +397,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 			},
 		},
@@ -414,13 +414,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					appender: newFakeAppender(conflictErrFn, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					appender: newFakeAppender(conflictErrFn, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					appender: newFakeAppender(conflictErrFn, commitErrFn, nil),
 				},
 			},
 		},
@@ -431,13 +431,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -448,13 +448,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -465,13 +465,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -482,13 +482,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -499,13 +499,13 @@ func TestReceiveQuorum(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -598,7 +598,7 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -609,7 +609,7 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -620,7 +620,7 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 			},
 		},
@@ -631,10 +631,10 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -645,13 +645,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -662,13 +662,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -679,13 +679,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -696,13 +696,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 			},
 		},
@@ -713,15 +713,15 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
+					appender:    newFakeAppender(nil, nil, nil),
 					appenderErr: appenderErrFn,
 				},
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
+					appender:    newFakeAppender(nil, nil, nil),
 					appenderErr: appenderErrFn,
 				},
 				{
-					appender:    newFakeAppender(nil, nil, nil, nil),
+					appender:    newFakeAppender(nil, nil, nil),
 					appenderErr: appenderErrFn,
 				},
 			},
@@ -733,13 +733,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 			},
 		},
@@ -750,13 +750,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					appender: newFakeAppender(conflictErrFn, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					appender: newFakeAppender(conflictErrFn, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, commitErrFn, nil),
+					appender: newFakeAppender(conflictErrFn, commitErrFn, nil),
 				},
 			},
 		},
@@ -767,13 +767,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -784,13 +784,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -801,13 +801,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil),
 				},
 				{
-					appender: newFakeAppender(conflictErrFn, nil, nil, nil),
+					appender: newFakeAppender(conflictErrFn, nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -818,13 +818,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil, nil),
+					appender: newFakeAppender(cycleErrors([]error{storage.ErrOutOfBounds, storage.ErrOutOfOrderSample, storage.ErrDuplicateSampleForTimestamp}), nil, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},
@@ -835,13 +835,13 @@ func TestReceiveWithConsistencyDelay(t *testing.T) {
 			wreq:              wreq1,
 			appendables: []*fakeAppendable{
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, commitErrFn, nil),
+					appender: newFakeAppender(nil, commitErrFn, nil),
 				},
 				{
-					appender: newFakeAppender(nil, nil, nil, nil),
+					appender: newFakeAppender(nil, nil, nil),
 				},
 			},
 		},