--- a/pkg/store/storepb/testutil/series.go
+++ b/pkg/store/storepb/testutil/series.go
@@ -77,11 +77,12 @@ func CreateHeadWithSeries(t testing.TB, j int, opts HeadGenOptions) (*tsdb.Head,
 	app := h.Appender(context.Background())
 	for i := 0; i < opts.Series; i++ {
 		ts := int64(j*opts.Series*opts.SamplesPerSeries + i*opts.SamplesPerSeries)
-		ref, err := app.Add(labels.FromStrings("foo", "bar", "i", fmt.Sprintf("%07d%s", ts, LabelLongSuffix)), ts, opts.Random.Float64())
+		ref, err := app.Append(0, labels.FromStrings("foo", "bar", "i", fmt.Sprintf("%07d%s", ts, LabelLongSuffix)), ts, opts.Random.Float64())
 		testutil.Ok(t, err)
 
 		for is := 1; is < opts.SamplesPerSeries; is++ {
-			testutil.Ok(t, app.AddFast(ref, ts+int64(is), opts.Random.Float64()))
+			_, err := app.Append(ref, nil, ts+int64(is), opts.Random.Float64())
+			testutil.Ok(t, err)
 		}
 	}
 	testutil.Ok(t, app.Commit())