… go.sum
… pkg/rules/manager_test.go
… pkg/store/bucket_test.go
… pkg/store/prometheus_test.go
… pkg/store/tsdb_test.go