@metalmatze 2021-03-16
Ran the new benchmark `BenchmarkMultiTSDB` (see PR) on both branches, main and this one.
```sh
 go test -bench=BenchmarkMultiTSDB -count=10 -benchtime=10000000x ./pkg/receive > /tmp/add.txt    #main
 go test -bench=BenchmarkMultiTSDB -count=10 -benchtime=10000000x ./pkg/receive > /tmp/append.txt #metalmatze:appender-append
```

Here are the benchstat results:
```
benchstat /tmp/add.txt /tmp/append.txt
name          old time/op    new time/op    delta
MultiTSDB-24     226ns ± 4%     219ns ± 4%  -2.85%  (p=0.003 n=10+9)

name          old alloc/op   new alloc/op   delta
MultiTSDB-24      171B ± 0%      171B ± 0%    ~     (all equal)

name          old allocs/op  new allocs/op  delta
MultiTSDB-24      0.00           0.00         ~     (all equal)
```

Looking at it from this benchmark nothing drastically changed, time got slightly better but probably normally not noticeable.