@metalmatze 2021-03-19
Yup, not WIP anymore. I applied your suggestion and we're ready to merge!